requirements.txt
django-allauth
celery
redis
django-apscheduler


